feat: ERC20AccessList first pass

+++ src/AccessListLib.sol
@@ -2,14 +2,12 @@
 pragma solidity ^0.8.16;
 import "forge-std/console.sol";
 // keccak256("access.list.library");
-bytes32 constant SLOT_ZERO =
-    0xf6b90c594052dd41030fa1082062d4d0952430a41949e8bea89c3f6ad3dc8a12;
+bytes32 constant SLOT_ZERO = 0xf6b90c594052dd41030fa1082062d4d0952430a41949e8bea89c3f6ad3dc8a12;
 
 // for maths
 uint256 constant UINT_SLOT_ZERO = uint256(SLOT_ZERO);
 
 library AccessListLib {
-
     function isWarm(bytes32 slot) internal returns (bool) {
         uint256 x;
         uint256 check1 = gasleft();
@@ -24,13 +22,16 @@ library AccessListLib {
         return check1 - check2 == check2 - check3;
     }
 
-    function addSlots(uint256 start, uint256 finish) internal returns (uint256 sum) {
+    function addSlots(uint256 start, uint256 finish)
+        internal
+        returns (uint256 sum)
+    {
         // TODO: Change to start and legnth
-        uint length = finish - start;
-        for (uint i = start; i < finish;) {
+        uint256 length = finish - start;
+        for (uint256 i = start; i < finish; ) {
             if (isWarm(slot(i))) {
                 // sum = sum + 0x1 << (finish - i - 1);
-                sum = sum + 2 ** (finish - i - 1);
+                sum = sum + 2**(finish - i - 1);
             }
             unchecked {
                 ++i;
@@ -53,7 +54,11 @@ library AccessListLib {
 
     function argAddress(uint256 index) internal returns (address) {
         // argIndex starts at 0 so first argument is 0, 2nd is 1 etc
-        return address(uint160(uint256(addSlots(32 + index * 256, 32 + index * 256 + 256))));
+        return
+            address(
+                uint160(
+                    uint256(addSlots(32 + index * 256, 32 + index * 256 + 256))
+                )
+            );
     }
-
 }
